Fix warnings related to SSC CSP project

+++ ssc/cmod_linear_fresnel_dsg_iph.cpp
@@ -239,12 +239,12 @@ public:
 		sim_setup.m_sim_time_start = 0.0;				//[s] starting first hour of year
 		sim_setup.m_sim_time_end = nhourssim*3600.0;	//[s] full year simulation
 
-		int steps_per_hour = 1;			//[-]
+		size_t steps_per_hour = 1;			//[-]
 
-		int n_wf_records = (int)weather_reader.m_weather_data_provider->nrecords();
+		size_t n_wf_records = weather_reader.m_weather_data_provider->nrecords();
 		steps_per_hour = n_wf_records / 8760;	//[-]
 
-		int n_steps_fixed = steps_per_hour*8760;	//[-]
+		size_t n_steps_fixed = steps_per_hour*8760;	//[-]
 		sim_setup.m_report_step = 3600.0 / (double)steps_per_hour;	//[s]
 		//***************************************************************************
 		//***************************************************************************
